Scripts/requirements.txt
pygame>=2.5.0
numpy>=1.21.0
PyYAML>=6.0
